Moved mock-dev.js and ui-dev.js to development folder.
