+++ development/ui-dev.js
@@ -17,10 +17,10 @@
 
 const render = require('react-dom').render
 const h = require('react-hyperscript')
-const Root = require('./ui/app/root')
-const configureStore = require('./development/uiStore')
-const states = require('./development/states')
-const Selector = require('./development/selector')
+const Root = require('../ui/app/root')
+const configureStore = require('./uiStore')
+const states = require('./states')
+const Selector = require('./selector')
 
 // logger
 const log = require('loglevel')
@@ -35,7 +35,7 @@ const firstState = states[selectedView]
 updateQueryParams(selectedView)
 
 // CSS
-const MetaMaskUiCss = require('./ui/css')
+const MetaMaskUiCss = require('../ui/css')
 const injectCss = require('inject-css')
